Update Spectrum for keyboard functionality.
Tab indexing on palette. Keyboard functionality of palette.
Applied aria-label, aria-pressed to the colour picker button. Other aria-labels to input. 
Have not been able to currently get keyboard functionality to work with the draggable elements. However this is purely for accessibility to the colourpicker. So the palette keyboard func should be good enough.